+++ src/core/image.rs
@@ -1,16 +1,19 @@
 use crate::core::colour_models::*;
 use crate::core::traits::PixelBound;
 use ndarray::prelude::*;
-use ndarray::s;
+use ndarray::{s, Data, DataMut, OwnedRepr, RawDataClone, ViewRepr};
 use num_traits::cast::{FromPrimitive, NumCast};
 use num_traits::Num;
-use std::marker::PhantomData;
+use std::{fmt, hash, marker::PhantomData};
+
+pub type Image<T, C> = ImageBase<OwnedRepr<T>, C>;
+pub type ImageView<'a, T, C> = ImageBase<ViewRepr<&'a T>, C>;
 
 /// Basic structure containing an image.
-#[derive(Clone, Eq, PartialEq, Hash, Debug)]
-pub struct Image<T, C>
+pub struct ImageBase<T, C>
 where
     C: ColourModel,
+    T: Data,
 {
     /// Images are always going to be 3D to handle rows, columns and colour
     /// channels
@@ -20,13 +23,14 @@ where
     /// number of channels in an image as this may cause other functionality to
     /// perform incorrectly. Use conversions to one of the `Generic` colour models
     /// instead.
-    pub data: Array3<T>,
+    pub data: ArrayBase<T, Ix3>,
     /// Representation of how colour is encoded in the image
     pub(crate) model: PhantomData<C>,
 }
 
-impl<T, C> Image<T, C>
+impl<T, U, C> ImageBase<U, C>
 where
+    U: Data<Elem = T>,
     T: Copy + Clone + FromPrimitive + Num + NumCast + PixelBound,
     C: ColourModel,
 {
@@ -43,7 +47,21 @@ where
             T2::from_f64(scaled).unwrap_or_else(T2::zero) + T2::min_pixel()
         };
         let data = self.data.map(rescale);
-        Image::<T2, C>::from_data(data)
+        Image::<_, C>::from_data(data)
+    }
+}
+
+impl<S, T, C> ImageBase<S, C>
+where
+    S: Data<Elem = T>,
+    T: Clone,
+    C: ColourModel,
+{
+    pub fn to_owned(&self) -> Image<T, C> {
+        Image {
+            data: self.data.to_owned(),
+            model: PhantomData,
+        }
     }
 }
 
@@ -56,7 +74,7 @@ where
     /// a colour model
     pub fn new(rows: usize, columns: usize) -> Self {
         Image {
-            data: Array3::<T>::zeros((rows, columns, C::channels())),
+            data: Array3::zeros((rows, columns, C::channels())),
             model: PhantomData,
         }
     }
@@ -65,7 +83,7 @@ where
     /// the data sizes don't match a zero filled image will be returned instead
     /// of panicking
     pub fn from_shape_data(rows: usize, cols: usize, data: Vec<T>) -> Self {
-        let data = Array3::<T>::from_shape_vec((rows, cols, C::channels()), data)
+        let data = Array3::from_shape_vec((rows, cols, C::channels()), data)
             .unwrap_or_else(|_| Array3::<T>::zeros((rows, cols, C::channels())));
 
         Image {
@@ -75,13 +93,28 @@ where
     }
 }
 
-impl<T, C> Image<T, C>
+impl<T, C> ImageBase<T, C>
 where
+    T: Data,
+    C: ColourModel,
+{
+    /// Create an image given an existing ndarray
+    pub fn from_array(data: ArrayBase<T, Ix3>) -> Self {
+        Self {
+            data,
+            model: PhantomData,
+        }
+    }
+}
+
+impl<T, U, C> ImageBase<T, C>
+where
+    T: Data<Elem = U>,
     C: ColourModel,
 {
     /// Construct the image from a given Array3
-    pub fn from_data(data: Array3<T>) -> Self {
-        Image {
+    pub fn from_data(data: ArrayBase<T, Ix3>) -> Self {
+        Self {
             data,
             model: PhantomData,
         }
@@ -101,13 +134,79 @@ where
     }
 
     /// Get a view of all colour channels at a pixels location
-    pub fn pixel(&self, row: usize, col: usize) -> ArrayView<T, Ix1> {
+    pub fn pixel(&self, row: usize, col: usize) -> ArrayView<U, Ix1> {
         self.data.slice(s![row, col, ..])
     }
 
+    pub fn into_type_raw<C2>(self) -> ImageBase<T, C2>
+    where
+        C2: ColourModel,
+    {
+        assert_eq!(C2::channels(), C::channels());
+        ImageBase::<T, C2>::from_data(self.data)
+    }
+}
+
+impl<T, U, C> ImageBase<T, C>
+where
+    T: DataMut<Elem = U>,
+    C: ColourModel,
+{
     /// Get a mutable view of a pixels colour channels given a location
-    pub fn pixel_mut(&mut self, row: usize, col: usize) -> ArrayViewMut<T, Ix1> {
+    pub fn pixel_mut(&mut self, row: usize, col: usize) -> ArrayViewMut<U, Ix1> {
         self.data.slice_mut(s![row, col, ..])
+    }
+}
+
+impl<T, U, C> fmt::Debug for ImageBase<U, C>
+where
+    U: Data<Elem = T>,
+    T: fmt::Debug,
+    C: ColourModel,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "ColourModel={:?} Data={:?}", self.model, self.data)?;
+        Ok(())
+    }
+}
+
+impl<T, U, C> PartialEq<ImageBase<U, C>> for ImageBase<U, C>
+where
+    U: Data<Elem = T>,
+    T: PartialEq,
+    C: ColourModel,
+{
+    fn eq(&self, other: &Self) -> bool {
+        self.model == other.model && self.data == other.data
+    }
+}
+
+impl<S, C> Clone for ImageBase<S, C>
+where
+    S: RawDataClone + Data,
+    C: ColourModel,
+{
+    fn clone(&self) -> Self {
+        Self {
+            data: self.data.clone(),
+            model: PhantomData,
+        }
+    }
+
+    fn clone_from(&mut self, other: &Self) {
+        self.data.clone_from(&other.data)
+    }
+}
+
+impl<'a, S, C> hash::Hash for ImageBase<S, C>
+where
+    S: Data,
+    S::Elem: hash::Hash,
+    C: ColourModel,
+{
+    fn hash<H: hash::Hasher>(&self, state: &mut H) {
+        self.model.hash(state);
+        self.data.hash(state);
     }
 }