+++ src/format/mod.rs
@@ -1,5 +1,6 @@
 use crate::core::traits::PixelBound;
 use crate::core::*;
+use ndarray::Data;
 use num_traits::cast::{FromPrimitive, NumCast};
 use num_traits::{Num, NumAssignOps};
 use std::fmt::Display;
@@ -8,17 +9,22 @@ use std::io::prelude::*;
 use std::path::Path;
 
 /// Trait for an image encoder
-pub trait Encoder<T, C>
+pub trait Encoder<T, U, C>
 where
+    U: Data<Elem = T>,
     T: Copy + Clone + Num + NumAssignOps + NumCast + PartialOrd + Display + PixelBound,
     C: ColourModel,
 {
     /// Encode an image into a sequence of bytes for the given format
-    fn encode(&self, image: &Image<T, C>) -> Vec<u8>;
+    fn encode(&self, image: &ImageBase<U, C>) -> Vec<u8>;
 
     /// Encode an image saving it to the file at filename. This function shouldn't
     /// add an extension preferring the user to do that instead.
-    fn encode_file<P: AsRef<Path>>(&self, image: &Image<T, C>, filename: P) -> std::io::Result<()> {
+    fn encode_file<P: AsRef<Path>>(
+        &self,
+        image: &ImageBase<U, C>,
+        filename: P,
+    ) -> std::io::Result<()> {
         let mut file = File::create(filename)?;
         file.write_all(&self.encode(image))?;
         Ok(())